refactor(service): 컴포넌트 폴더구조 변경 (도메인 별로)
