+++ apps/service/src/components/common/Buttons/SolveButton.tsx
